requirements.txt
aenum==3.1.15
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiohttp-retry==2.9.1
aiosignal==1.3.2
abydos==0.5.0
auth0-python==4.9.0
annotated-types==0.7.0
annoy==1.17.3
anyio==4.9.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
audioop-lts==0.2.1
av==14.2.0
azure-cognitiveservices-speech==1.42.0
azure-core
azure-identity>=1.16.0
msal==1.32.3
msal-extensions==1.3.1
babel==2.16.0
backoff==2.2.1
boto3==1.38.39
botocore==1.38.39
cachetools==5.5.2
cartesia==1.3.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
cryptography==45.0.4
dataclasses-json==0.6.7
deepgram-sdk==3.10.1
deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
docopt==0.6.2
docstring-parser==0.16
elevenlabs==1.52.0
eval-type-backport==0.2.2
exceptiongroup==1.2.2
expiringdict==1.2.2
fastapi==0.115.4
ffmpeg-python==0.2.0
ffmpy==0.5.0
filelock==3.18.0
flatbuffers==25.2.10
freezegun==1.5.1
frozenlist==1.5.0
fsspec==2025.3.0
future==1.0.0
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-secret-manager==2.23.2
googleapis-common-protos==1.69.2
gotrue==2.11.2
gradio==5.23.1
gradio-client==1.8.0
groovy==0.1.2
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.62.3
gtts==2.5.4
h11==0.14.0
h2==4.1.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib-metadata==8.5.0
iniconfig==2.0.0
iterators==0.2.0
kubernetes==32.0.1
jellyfish==1.1.3
jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.16
langchain-aws==0.2.26
langchain-core==0.3.66
langchain-openai==0.3.11
langchain-text-splitters==0.3.5
langfuse==2.58.1
langgraph==0.3.21
langgraph-checkpoint==2.0.10
langgraph-prebuilt==0.1.7
langgraph-sdk==0.1.51
langsmith==0.3.1
livekit==1.0.0
livekit-api==1.0.2
markupsafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.2.0
mypy-extensions==1.0.0
networkx==3.4.2
nltk==3.9.1
num2words==0.5.13
numpy==2.2.4
onnxruntime==1.21.0
openai==1.69.0
opentelemetry-api==1.29.0
opentelemetry-distro==0.50b0
opentelemetry-exporter-otlp==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-requests==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
orjson==3.10.15
ormsgpack==1.9.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pluggy==1.5.0
postgrest==0.19.3
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.30.2
psutil==7.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.1
pydantic-core==2.33.0
pydub==0.25.1
pyjwt==2.10.1
pyRFC3339==2.0.1
pytest==8.3.5
pytest-mock==3.14.0
pytest-asyncio==0.26.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2024.2
pywav==1.1
pyyaml==6.0.2
questionary==2.1.0
realtime==2.2.0
redis==4.5.5
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
s3transfer==0.7.0
safehttpx==0.1.6
safetensors==0.5.3
scipy==1.15.2
semantic-version==2.10.0
semver==3.0.4
setuptools==78.1.0
silero-vad==5.1.2
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.13.1
sqlalchemy==2.0.37
starlette==0.41.3
storage3==0.11.1
strenum==0.4.15
supabase==2.12.0
supafunc==0.9.2
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
termcolor==2.5.0
text2num==2.5.1
tiktoken==0.8.0
tokenizers==0.21.1
tomlkit==0.13.2
torch==2.6.0
torchaudio==2.6.0
tqdm==4.67.1
transformers==4.50.3
twilio==9.4.4
typer==0.15.2
types-protobuf==4.25.0.20240417
typing-extensions==4.13.0
typing-inspect==0.9.0
typing-inspection==0.4.0
tzdata==2025.1
urllib3==2.0.7
uv==0.5.29
uvicorn==0.21.1
watchfiles==1.0.4
websockets==13.1
wordninja==2.0.0
wordsegment==1.3.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
prompt-toolkit==3.0.50
wcwidth==0.2.13

durationpy==0.10
oauthlib==3.2.2
requests-oauthlib==2.0.0
websocket-client==1.8.0
resend==2.10.0
py4j==0.10.9.9